refactor: 파일명 변경 (content -> contents)
