+++ src/views/main/components/quiz-banner-contents.tsx
